tests/robustness: Refactor patch operation functions
